+++ addon/objects/uml-primitives/links-view/fd-empty-view.js
@@ -263,7 +263,7 @@ export let EmptyView = joint.dia.LinkView.extend({
     this._setVerticesValue();
     let readonly = this.paper.options.interactive;
     if (readonly && typeof readonly === 'object') {
-      $(this.paper.el).find('input,textarea').addClass('click-disabled');
+      this.$box.find('input,textarea').addClass('click-disabled');
     }
 
     let coordinates = forPointerMethodOverrideResizeAndDnd(evt, x, y);
@@ -278,8 +278,9 @@ export let EmptyView = joint.dia.LinkView.extend({
   pointerup(evt, x, y) {
     this._checkVerticesChanges();
     let readonly = this.paper.options.interactive;
+
     if (readonly && typeof readonly === 'object') {
-      $(this.paper.el).find('input,textarea').removeClass('click-disabled');
+      this.$box.find('input,textarea').removeClass('click-disabled');
     }
 
     let coordinates = forPointerMethodOverrideResizeAndDnd(evt, x, y);