+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Ember Flexberry Designer Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased] - 2026-03-11
+## [Unreleased]
 ### Fixed
+* Fixed diagram class edit
+* Fixed logo shrink
 * Changed card on generation page
 
 ## [1.1.2-beta.7] - 2026-03-05